+++ browser/VencordNativeStub.ts
@@ -45,7 +45,7 @@ window.VencordNative = {
         deleteTheme: (fileName: string) => DataStore.del(fileName, themeStore),
         getThemesDir: async () => "",
         getThemesList: () => DataStore.entries(themeStore).then(entries =>
-            entries.map(([name, css]) => ({ fileName: name as string, content: css }))
+            entries.map(([name, css]) => getThemeInfo(css, name.toString()))
         ),
         getThemeData: (fileName: string) => DataStore.get(fileName, themeStore),
         getSystemValues: async () => ({}),
