--- patches/@types__less@3.0.8.patch
@@ -1,13 +1,0 @@
-diff --git a/index.d.ts b/index.d.ts
-index eb4f07d47b932fb9cc8c8cd451ab107f648bd013..18a3e15a1997734e1773718e5be55d252ed9478c 100644
---- a/index.d.ts
-+++ b/index.d.ts
-@@ -306,7 +306,5 @@ interface LessStatic {
- }
- 
- declare module "less" {
--    export = less;
-+    export = LessStatic;
- }
--
--declare var less: LessStatic;
