--- src/components/ThemeSettings/components/colorStyles.css
@@ -1,19 +1,0 @@
-.vc-usercss-settings-color-swatch-row {
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    align-items: center;
-}
-
-.vc-usercss-settings-color-swatch-row > span {
-    display: block;
-    flex: 1;
-    overflow: hidden;
-    margin-top: 0;
-    margin-bottom: 0;
-    color: var(--text-strong);
-    line-height: 24px;
-    font-size: 16px;
-    font-weight: 500;
-    overflow-wrap: break-word;
-}
