--- src/components/ThemeSettings/components/index.ts
@@ -1,12 +1,0 @@
-/*
- * Vencord, a Discord client mod
- * Copyright (c) 2023 Vendicated and contributors
- * SPDX-License-Identifier: GPL-3.0-or-later
- */
-
-export * from "./SettingBooleanComponent";
-export * from "./SettingColorComponent";
-export * from "./SettingNumberComponent";
-export * from "./SettingRangeComponent";
-export * from "./SettingSelectComponent";
-export * from "./SettingTextComponent";
