--- src/components/ThemeSettings/components/styles.css
@@ -1,3 +1,0 @@
-.vc-settings-boolean-switch {
-    margin-bottom: 0.5em;
-}
