+++ src/components/settings/tabs/index.ts
@@ -6,7 +6,6 @@
 
 import "./styles.css";
 
-export { default as ThemesTab } from "../../ThemeSettings/ThemesTab";
 export * from "./BaseTab";
 export { default as ChangelogTab } from "./changelog";
 export { default as PatchHelperTab } from "./patchHelper";
@@ -15,5 +14,6 @@ export { openContributorModal } from "./plugins/ContributorModal";
 export { openPluginModal } from "./plugins/PluginModal";
 export { default as BackupAndRestoreTab } from "./sync/BackupAndRestoreTab";
 export { default as CloudTab } from "./sync/CloudTab";
+export { default as ThemesTab } from "./themes";
 export { default as UpdaterTab } from "./updater";
 export { default as VencordTab } from "./vencord";
