--- src/utils/themes/index.ts
@@ -1,17 +1,0 @@
-/*
- * Vencord, a Discord client mod
- * Copyright (c) 2023 Vendicated and contributors
- * SPDX-License-Identifier: GPL-3.0-or-later
- */
-
-import type { UserstyleHeader } from "usercss-meta";
-
-import type { UserThemeHeader } from "./bd";
-
-export type ThemeHeader = {
-    type: "other";
-    header: UserThemeHeader;
-} | {
-    type: "usercss";
-    header: UserstyleHeader;
-};
